requirements.txt
black>=23.3.0,<23.4
celery[redis]>=5.3.0,<5.4
currencyapicom>=0.1.1,<0.2
dj-database-url>=2.0.0,<2.1
Django>=4.2.2,<4.3
djangorestframework>=3.14.0,<3.15
django-celery-beat>=2.5.0,<2.6
djangorestframework-api-key>=2.3.0,<2.4
flake8>=6.0.0,<6.1
gunicorn>=20.1.0,<20.2
psycopg2-binary>=2.9.6,<3.0